2차 세션만료 수정(임시)

+++ app.py
@@ -131,11 +131,14 @@ def home():
 def home():
     global something_doing
     global login_check
+
     if not login_check:
+        something_doing = False
+        print('something_doing = False : 137')
         return redirect(url_for('login'))
 
     something_doing = True
-    print('something_doing = True')
+    print('something_doing = True : 141')
     model = '입력...'
     cloth_type = 'Mbtm'
     MSRP = '조회시 출력'
@@ -143,12 +146,11 @@ def home():
     cloth_type_original = 'Mbtm'
     invalidity = 'false'
     Image_num = 0
-    size = {'mbtm' : ['28','30','32','34','36','38','40','계'],
-            'wbtm' : ['23/30','24/31','25','26','27','28','29','계'],
-            'mtop' : ['85(XS)','90(S)','95(M)','100(L)','105(XL)','X','X','계'],
-            'wtop' : ['80(XS)','85(S)','90(M)','95(L)','X','X','X','계'],
-            'acc'  : ['70','75','90','95','100/F','X','X','계']}
-    
+    size = {'mbtm' : ['28'     ,'30'   ,'32'   ,'34'    ,'36'     ,'38','40','계'],
+            'wbtm' : ['23/30'  ,'24/31','25'   ,'26'    ,'27'     ,'28','29','계'],
+            'mtop' : ['85(XS)' ,'90(S)','95(M)','100(L)','105(XL)','X' ,'X' ,'계'],
+            'wtop' : ['80(XS)' ,'85(S)','90(M)','95(L)' ,'X'      ,'X' ,'X' ,'계'],
+            'acc'  : ['70'     ,'75'   ,'90'   ,'95'    ,'100/F'  ,'X' ,'X' ,'계']}
     
     initial_data = {'size' : ["", "", "", "", "", "", "", ""],
                     'NC불광': ["", "", "", "", "", "", "", ""],
@@ -186,6 +188,8 @@ def home():
                 if model == '':
                     model = '입력...'
 
+                something_doing = False
+                print('something_doing = False : 192')
                 return render_template('home.html', stock_data = stock_data, sell_data = sell_data, MSRP = MSRP, Season = Season, model = model, cloth_type = cloth_type_original, invalidity = invalidity, ware_data = ware_data)
             
             #재고 현황 파악
@@ -219,7 +223,8 @@ def home():
             Season = getSeason(workbook, model, cloth_type)
         else:
             model = '입력...'
-
+    something_doing = False
+    print('something_doing = False : 227')
     return render_template('home.html', stock_data = stock_data, sell_data = sell_data, MSRP = MSRP, Season = Season, model = model, cloth_type = cloth_type_original, invalidity = invalidity, ware_data = ware_data, Image_num = Image_num)
 
 
@@ -257,6 +262,7 @@ def login():
     global login_check
 
     something_doing = True
+    print('something_doing = True : 265')
     
     id = ""
     pw = ""
@@ -269,15 +275,20 @@ def login():
                 login_check = True
 
                 maintain_session = True
+                something_doing = False
+                print('something_doing = False : 279')
                 return redirect(url_for('home'))
             if id == 'jinha12345':
                 login_check = True
 
                 maintain_session = True
+                something_doing = False
+                print('something_doing = False : 286')
                 return redirect(url_for('home'))
             else:
                 flash('아이디 또는 비밀번호가 일치하지 않습니다.', 'error')
-    
+    something_doing = False
+    print('something_doing = False : 291')
     return render_template('login.html')
 
 '''
@@ -305,6 +316,8 @@ def passwordgen():
 
 @app.route('/passwordgen', methods = [ "GET", "POST"])
 def passwordgen():
+    global something_doing
+    something_doing = True
     id = ""
     pw = ""
     new_id = ""
@@ -319,52 +332,56 @@ def passwordgen():
             if id == 'admin' and pw == 'admin':
                 login_check = True
                 new_pw = passgen(new_id)
-
+                something_doing = False
+                print('something_doing = False : 336')
                 return render_template('passwordgen.html', id = id, pw = pw, new_id = new_id, new_pw = new_pw)
-    
+    something_doing = False
+    print('something_doing = False : 339')
     return render_template('passwordgen.html')
 
 
 @socketio.on('disconnect')
 def handle_disconnect():
-    #global pong
+    global pong
     global something_doing
-    #pong = False
+    pong = False
     print('Client disconnected')
 
     if something_doing:
         print('But something doing...')
         something_doing = False
+        print('something_doing = False : 353')
         return
 
     # 연결이 끊겼을 때 타이머 시작
     disconnect_timer = 5
-    while disconnect_timer > 0:# and pong == False:
+    while disconnect_timer > 0 and pong == False:
         if something_doing:
             print('But something doing...')
             something_doing = False
-            return
+            print('something_doing = False : 362')
+            return None
         #print(f"Waiting for pong... {disconnect_timer} seconds left")
         print(f"Waiting for something doing... {disconnect_timer} seconds left")
         disconnect_timer -= 1
-        #socketio.emit('ping')  # 클라이언트에게 ping 이벤트를 보냄
+        socketio.emit('ping')  # 클라이언트에게 ping 이벤트를 보냄
         time.sleep(1)
     else:
         #if pong == False:
-        if not something_doing:
+        if not something_doing and pong == False:
             #print('No pong received, considering the client as disconnected')
             print('Doing nothing, considering the client as disconnected')
             os.kill(os.getpid(), signal.SIGINT)
         # 여기에서 페이지 종료에 따른 추가 로직을 수행할 수 있음
-    #pong = False
+    pong = False
 
-'''
+
 @socketio.on('pong')
 def handle_pong():
     global pong
     print('Pong received from the client')
     pong = True
-'''
+
 
 if __name__ == '__main__':
     #JsonKey를 drive, temp, appdata 동기화
